Refactor: 예외 발생시 재입력 기능 단일화
- 입력부에서 예외처리의 재입력 기능이 공통적으로 발생하기에 이를 한개의 메서드로 처리함.
